chore(examples): simplify `multiple`, add `event_options`
